[refactor] 파일이름 수정 issue-number
- globalSeletor => globalSelector
