+++ front-end/src/components/Lobby/LobbyButtons.tsx
@@ -10,7 +10,7 @@ import VerfiyPasswordModal from './VerifyPasswordModal';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import globalAtom from '../../recoilStore/globalAtom';
 import { modalPropsType } from '../public/Modal';
-import globalSelector from '../../recoilStore/globalSeletor';
+import globalSelector from '../../recoilStore/globalSelector';
 
 interface roomInterface {
   [prop: string]: any;
